Enhanced Bot Processing View
